# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "srvcs-alltrue",
    "description": "",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [],
        "summary": "`GET /` — service identity (srvcs service standard).",
        "operationId": "index",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [],
        "summary": "`POST /` — are all values in the list `true`?",
        "description": "This service does no logic of its own. It folds the list through\n`srvcs-and`, starting from `true`: `acc = and(acc, v)` for each element. The\nconjunction of the empty list is `true` and makes no dependency calls. If\n`and` rejects an element the `422` is forwarded; if `and` is unreachable this\nservice reports itself degraded rather than guessing.",
        "operationId": "evaluate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllTrueResponse"
                }
              }
            }
          },
          "422": {
            "description": "an element is not a valid boolean (forwarded from srvcs-and)"
          },
          "500": {
            "description": "srvcs-and returned an unusable response"
          },
          "503": {
            "description": "the srvcs-and dependency is unavailable"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AllTrueResponse": {
        "type": "object",
        "required": [
          "values",
          "result"
        ],
        "properties": {
          "result": {
            "type": "boolean"
          },
          "values": {
            "type": "object"
          }
        }
      },
      "EvalRequest": {
        "type": "object",
        "required": [
          "values"
        ],
        "properties": {
          "values": {
            "type": "object",
            "description": "The list of booleans to AND together. An empty list is `true`."
          }
        }
      },
      "Info": {
        "type": "object",
        "required": [
          "service",
          "concern",
          "depends_on"
        ],
        "properties": {
          "concern": {
            "type": "string"
          },
          "depends_on": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "service": {
            "type": "string"
          }
        }
      }
    }
  }
}